% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_number}
\alias{cluster_number}
\title{Clustering number}
\usage{
cluster_number(data, method = "elbow", max_clusters = 50)
}
\arguments{
\item{data}{Matrix, cleaned}

\item{method}{Character, choose c(elbow, gapstats)}
}
\value{
a plot to understand how maany k you should use
}
\description{
prima funzion eper esploazione e grazie al grafico capire quanti k passare
seconda effettiva per calcorare i cluster con i metodi
}
\details{
TODO: -Aggiungi examples nelle documentazioni
}
