% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering method}
\usage{
clustering(
  data,
  method = "kmeans",
  metric = NULL,
  centers = NULL,
  eps = NULL,
  minpts = NULL
)
}
\arguments{
\item{data}{Matrix or data.frame, cleaned}

\item{method}{Character, method = c("kmeans", "hierarchical", "kmedoids", "DBSCAN")}

\item{metric}{Character, metric for kmedoids ('euclidean','manhattan')}

\item{centers}{Numerical, number of clusters}

\item{eps}{Numerical, for DBSCAN}

\item{minpts}{Numerical, for DBSCAN}
}
\value{
A list with interactive plot, PCA result and clustering info
}
\description{
take in input a clustering method and gives a list of useful result
}
