% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataVisualiztion.R
\name{data_visualization}
\alias{data_visualization}
\title{Simple viewer for clustering results}
\usage{
data_visualization(clustering_result, what = "plot")
}
\arguments{
\item{clustering_result}{List, returned by clustering_method()/cluster_number()}

\item{what}{Character, choose "plot", "pca", "summary", "dendrogram"}
}
\value{
Console output or plot
}
\description{
Takes the results from clustering function and helps to retrieve information
}
