% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploraAnalysis.R
\name{explorative_analysis}
\alias{explorative_analysis}
\title{Explorative analysis for preprocessed data}
\usage{
explorative_analysis(
  data,
  analysis_type = "CorrHeatmap",
  gene1 = NULL,
  gene2 = NULL
)
}
\arguments{
\item{data}{(preprocessed) Matrix, ready for the analysis}

\item{analysis_type}{Character, c("corrHeatmap", "pca", "density", lm"(metti anche qq), "scatter")}

\item{gene1}{Character, primo gene per confronto}

\item{gene2}{Character, gene 2 per confronto}
}
\value{
description of the varoiuos output
}
\description{
Given a preprocessed df return an explorative matrix
-heatmap
-lm model?
-par(mfrow=c(2,2)) magari con qq e piu grafici
- scatter o violin
}
