% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.R
\name{gsea_analysis}
\alias{gsea_analysis}
\title{Gene set enrichment analysis}
\usage{
gsea_analysis(data, keyType = "ENSEMBL", condition = "no", groups = NULL)
}
\arguments{
\item{data}{Matrix or Data Frame, with gene expression (genes in rows)}

\item{keyType}{Character, gene ID type (e.g. 'ENSEMBL', 'SYMBOL')}

\item{condition}{Character, 'yes' or 'no', indicates if groups are used}

\item{groups}{Vector of conditions, required if condition = 'yes'}
}
\value{
A list with GSEA result and plot (or NULL plot if empty result)
}
\description{
Perform a gene set enrichment analysis of the input data
}
