% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessData.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess DNA expression data}
\usage{
preprocess(
  data,
  normalize = TRUE,
  method = "log",
  batch_effect = FALSE,
  batch = NULL
)
}
\arguments{
\item{data}{Matrix or ExpressionSet, of expression values}

\item{normalize}{Logical, decide if you want to normalize your data, default=TRUE}

\item{method}{Character, choose the method to normalize between ('log', 'quantile'), default='log'}

\item{batch_effect}{Logical, decide whether to apply batch effect correction, default=FALSE}

\item{batch}{Vector, indicating the batch assignment, not required for ExpressionSet}
}
\value{
A normalized matrix or ExpressionSet
}
\description{
Given genes expression data (as matrix or ExpressionSet), preprocess giving a normalized df
}
