% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{cluster_number}
\alias{cluster_number}
\title{Clustering number}
\usage{
cluster_number(data, method = "elbow", max_clusters = 50)
}
\arguments{
\item{data}{Matrix, cleaned and preprocessed gene expression or feature data}

\item{method}{Character, choose c(elbow, gapstats) estimation method}

\item{max_clusters}{Integer, maximum number of clusters to consider (default = 50)}
}
\value{
a plot to understand how maany k you should use
}
\description{
This function supports two techniques: elbow method and gap statistics. To
help estimate the optimal number of clusters to use in downstream clustering.
}
\examples{
# Simulated example
set.seed(42)
mat <- matrix(rnorm(100 * 10), nrow = 100, ncol = 10)

# Elbow method
cluster_number(mat, method = "elbow")

# Gap statistic
cluster_number(mat, method = "gapstats")
}
