% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clustering}
\alias{clustering}
\title{Clustering method}
\usage{
clustering(
  data,
  method = "kmeans",
  metric = NULL,
  centers = NULL,
  eps = NULL,
  minpts = NULL
)
}
\arguments{
\item{data}{Matrix or data.frame, cleaned and normalized}

\item{method}{Character, method = c("kmeans", "hierarchical", "kmedoids", "DBSCAN")}

\item{metric}{Character, distance metric for kmedoids ('euclidean','manhattan')}

\item{centers}{Numerical, number of clusters}

\item{eps}{Numerical, epsilon value for DBSCAN}

\item{minpts}{Numerical, minimum points for DBSCAN}
}
\value{
A list with interactive plot, PCA result and clustering info
}
\description{
This function performs clustering using different techniques
(k-means, k-medoids,hierarchical, DBSCAN) and returns outputs
including an interactive PCA plot,clustering assignments, and summary.
}
\examples{
# Example using the iris dataset (excluding species label)
data(iris)
iris_data <- iris[, -5]  # Remove species label

# K-means clustering
result_kmeans <- clustering(data = iris_data, method = "kmeans", centers = 3)
print(result_kmeans$cluster_information)

# K-medoids clustering
result_kmedoids <- clustering(data = iris_data, method = "kmedoids", centers = 3, metric = "euclidean")
print(result_kmedoids$cluster_information)

# Hierarchical clustering
result_hierarchical <- clustering(data = iris_data, method = "hierarchical", centers = 3)
print(result_hierarchical$cluster_information)

# DBSCAN clustering
result_dbscan <- clustering(data = iris_data, method = "DBSCAN", eps = 0.5, minpts = 5)
print(result_dbscan$dbscan)
}
