% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploraAnalysis.R
\name{explorative_analysis}
\alias{explorative_analysis}
\title{Explorative analysis for preprocessed data}
\usage{
explorative_analysis(
  data,
  analysis_type = "CorrHeatmap",
  gene1 = NULL,
  gene2 = NULL
)
}
\arguments{
\item{data}{(preprocessed) Matrix, ready for the analysis}

\item{analysis_type}{Character, c("corrHeatmap", "pca", "density", lm"(metti anche qq), "scatter")}

\item{gene1}{Character, first gene for comparison}

\item{gene2}{Character, second gene for comparison}
}
\value{
description of the varoiuos output
}
\description{
This function performs different types of exploratory analysis on a preprocessed
gene expression matrix. Supported types include correlation heatmaps, PCA,
gene expression scatterplots, Q-Q plots and histograms.
}
\examples{
mat <- matrix(rnorm(1000), nrow = 10, dimnames = list(paste0("Gene", 1:10), paste0("Sample", 1:100)))
explorative_analysis(mat, analysis_type = "CorrHeatmap")
explorative_analysis(mat, analysis_type = "pca")
explorative_analysis(mat, analysis_type = "scatter", gene1 = "Gene1", gene2 = "Gene2")
explorative_analysis(mat, analysis_type = "lm", gene1 = "Gene1", gene2 = "Gene2")
}
