% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreProcessData.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess DNA expression data}
\usage{
preprocess(
  data,
  normalize = TRUE,
  method = "log",
  batch_effect = FALSE,
  batch = NULL,
  fallback_n = 5
)
}
\arguments{
\item{data}{Matrix, expression values}

\item{normalize}{Logical, decide if you want to normalize your data, default=TRUE}

\item{method}{Character, choose the method to normalize between ('log', 'quantile'), default='log'}

\item{batch_effect}{Logical, decide whether to apply batch effect correction, default=FALSE}

\item{batch}{Vector, indicating the batch assignment, required for batch effect correction}

\item{fallback_n}{Integer, number of random genes to retain if all get removed, default=5}
}
\value{
A normalized matrix
}
\description{
Given genes expression data (as matrix), preprocess with different
normalization methods giving as output normalized data
}
\examples{
# Simulate a matrix
set.seed(123)
expr_data <- matrix(abs(rnorm(600, mean = 10, sd = 3)), nrow = 100, ncol = 6)
colnames(expr_data) <- paste0("Sample", 1:6)

# Default preprocessing with log2 normalization
norm_data <- preprocess(expr_data)

# Preprocessing with quantile normalization
norm_data_q <- preprocess(expr_data, method = "quantile")

# Preprocessing with batch effect (2 groups of 3 samples each)
batch <- c(1,1,1,2,2,2)
norm_batch_data <- preprocess(expr_data, batch_effect = TRUE, batch = batch)
}
